Milestone/0.1.3 (#66)
* Implementation of #23 Stropping and #62 Add globals

This is a complete PR for the Stropping implementation and the library
access to add globals with some other work on #64 mixed in. It is not
a 1.1.3 release candidate until #64 is complete.

* Fixing tests

1. rebuilt docker to get latest py38
2. Fixing use of python keyword.
3. Making c reserved pattern more accurate.

* completion of #64, implicit language support.

* Refactor to support completing #64 properly.

This change moves language support up out of the jinja package. By
making language support template engine agnostic we prevent spreading
Jinja2 types throughout the namespace handling code et-al. This will
allow us to reuse the encoding and stropping logic on the namespace
generator.

* The real fix for #64

Let's try this again, shall we?

* Update src/nunavut/lang/py.py

Co-Authored-By: Pavel Kirienko <pavel.kirienko@gmail.com>

* Update src/nunavut/lang/py.py

Co-Authored-By: Pavel Kirienko <pavel.kirienko@gmail.com>