lang/c: migrate to 'static inline' from 'inline' functions
The problem with using `inline` functions is that functions had to be
declared `extern inline` before they could be used, which is a fairly
major usability issue. You would expect to include a header and then
immediately be able to use the functions.

By declaring these functions `static inline`, we can now use these
functions anywhere inside the compilation unit.

The other solution was to declare the functions in the header and
generate seperate source files defining them, however that adds extra
build complexity, without a huge amount of benefit. However the
likelihood of these functions being included in multiple compilation
units when code size is a major concern is fairly small.