Changing bit-length type
1. fixing buildkite artifact upload so the HTML coverage reports will
show up.
2. Formatting fixes in serialization.j2
3. Adding a unit test for nunavutInternalGetBitCopySize
4. Although I didn't want to use size_t for bit-lenth values the
alternatives all had performance impacts where the only argument against
size_t is convention. The only other viable option was uintptr_t but
that suffers from the same problem as size_t; the type is sized for a
platform based on byte lengths. There simply isn't a bit-length type in
C or C++ so we'll never get this right without defining our own and we
don't want to do that.