Add filter for selective blank lines removal (#151)
* Add filter for selective blank lines removal

(╯°□°)╯︵ ┻━┻

Humans should not do things that can be trivially automated away.
Yes, Jinja allows one to do precise whitespace control.
It's a nice feature that works great if you prefer tinkering with whitepace management rather than focusing on the business logic.
I want just describe the logic and let something else format the code for me.

* Bump the version number

* List the new filer in the docs.

This is not DRY though. One could automate this with sphinx-computron. It is also used in PyDSDSL and PyUAVCAN.

(☞ﾟヮﾟ)☞ ┻━┻