Long-overdue refactor for python core
A lot of stuff got added without a good refactor to DRY out the library
and tighten up the testing. This change  brings HTML into the fold and
prepairs the way for standards variations in templates (e.g. C++11
versus C++17, etc).

The docs were tightened up a bit and the filter/test loading mechanism
was completely rewritten.

A couple of bugs were found and fixed.