Fixing up the "uuid" thing in html
This fixes python 3.5 but also I found out why the unique function was
working even after we re-enabled caching. I accidentally did a really
intelligent thing in adding a special annotation to the unique filter
that has the side-effect of making that part of the template volatile.
So I documented this behaviour and am pretending I'm just that smart.