Long-overdue refactor for python core (#212)
* Long-overdue refactor for python core

A lot of stuff got added without a good refactor to DRY out the library
and tighten up the testing. This change  brings HTML into the fold and
prepairs the way for standards variations in templates (e.g. C++11
versus C++17, etc).

The docs were tightened up a bit and the filter/test loading mechanism
was completely rewritten.

A couple of bugs were found and fixed.

* fixing linter and typo

* Update src/nunavut/lang/cpp/templates/_fields_as_variant.j2

Co-authored-by: Pavel Kirienko <pavel.kirienko@gmail.com>

* fixing classifiers in setup.cfg

* using fields_except_padding

* adding generated api doc back in

* Fixing up the "uuid" thing in html

This fixes python 3.5 but also I found out why the unique function was
working even after we re-enabled caching. I accidentally did a really
intelligent thing in adding a special annotation to the unique filter
that has the side-effect of making that part of the template volatile.
So I documented this behaviour and am pretending I'm just that smart.

* Removing redundant doctest that was not supportable

Also ended up cleaning the conftest up a bit.

* adding tag to variant

Co-authored-by: Pavel Kirienko <pavel.kirienko@gmail.com>