First iteration on v2
Nunavut V2 is not a huge change from v1 but it the Python APIs are
different enough to warrent the semantic, major revision bump.

The core difference is a refactor of the Language objects to fully
support the properties.yaml schema and allow overrides to be provided to
nnvg.

A follow-up to this change, before v2 is published, will be a refactor
of the internal package and module names to improve encapsulation and to
be more consistent with the style of other OpenCyphal python projects.