Moving to internal VLA type
This release changes the default C++ generator to email a built-in VariableLengthArray type. This type makes no use of the standard library's containers and allocators and is optimized for Cyphal types. It does support using STL allocators but also includes a special pattern that allows the VLA to be backed by a fixed array with minimal overhead.


# PR activity

* Fix missing array capacity constant generated

* Fix typo

* Fix typo[2]

* Update src/nunavut/lang/cpp/templates/_fields.j2

Merge suggestion by @pavel-kirienko

Co-authored-by: Pavel Kirienko <pavel.kirienko@gmail.com>

* Proper support for variable_length_array

This change fixes the support for a built-in, default variable length array implementation for C++ types that can be overridden by the user. To accomplish this the concept of “support resource type” was added to allow different behavior for serialization support resource than for datatype support resources.

* fixing windows build

* checkpoint

almost there.

* Carefully adding a couple more features

* it aint pretty but it works

* not sure why you no Windows?

* fixing windows and adding except specifiers to new contrustors

* fixing test

* fixing gcc builds

* fixing logic error in fp test

* tests are broken but pushing to save my work

* fixing test

* fixing 32bit builds

* removing vestigal comment

* fixing linter error

* fixing code smell

Co-authored-by: Pavel Pletenev <cpp.create@gmail.com>
Co-authored-by: Pavel Kirienko <pavel.kirienko@gmail.com>