Fix for #277
This fixes, documents, and covers with test the ability to use std::vector instead of the built-in variable-length array type.

Also getting a head-start on #224

Also,also got rid of rstcheck as that project seems to be a hot mess. Too many false positives and major bugs not getting fixed. We’ll just have to manually test the docs using the tox -e docs target.