lang\c\templates\serialization.j2 seems working
Work of serialization.j2 was tested visually only. Not automatic

It is proposed to escape of the term "byte", since it is understood in different ways. Use "octet" and "char" instead. "Octet" is always 8 bits and is used when aligning DSDL messages. The size of the "char" can be 8, 16, 32 bits (CHAR_BIT), depending on the compiler