Nunavut 3.0 preview
This is a major change to the internal structure of Nunavut and the nnvg
command-line. At this level it is a breaking change although the code
generated for c has not changed (nor has experimental language support
but these are not as rigorously tested). This change goes along with
pydsdl 1.22+ which enables building types by a list of dsdl files and
generating the dependencies each target file requires.

The following fixes and features are included in this preview:

1. Meta-data in generated headers no longer includes absolute paths nor
   date/time information making the generation passes reproducible.
   a command-line option `--embed-auditing-info` was added to restore
   the previous behaviour.

2. Support for the CYPHAL_PATH environment variable was added.

3. All paths can be specified as `/path/to/root:nested/Type.0.1.dsdl`
   where the colon separates the path to the root namespace and the part
   after the colon is the path from this root to a DSDL file.

4. The internal architecture of the Nunavut python library has been
   rationalized considerably with a clean separation between CLI logic
   and the library interface. Major new layers include the redesigned
   namespace and generator modules.

5. The concept of "namespace index" was added although the
   implementation hasn't been piped through to the code generators yet.
   A future update will utilize this new concept to support multi-root
   generator templates that can generate files like GCC dep files,
   single-headers with all types included, and HTML index documents.
   Much of the internal restructuring was in support of this new
   concept.

This is a preview release and, as such, normal semantic versioning rules
are suspended until an official release has been made.